Clean up code on MYTrips.ysx
